pyproject.toml
[tool.poetry]
name = "rollin"
version = "0.1.0"
description = "This is my Trading Bot"
authors = ["Sanchay Jain"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiofiles = "^24.1.0"
uvloop = "^0.21.0"
websockets = "^14.0"
ta-lib = "^0.4.32"
django = "^5.1.3"
psycopg = "^3.2.3"
python-dotenv = "^1.0.1"
python-binance = "1.0.19"
pandas = "^2.2.3"
loguru = "^0.7.3"
djangorestframework = "^3.15.2"
django-browser-reload = "^1.17.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
